Rfc: 위치 정보 허가 관련 hook 생성 후 파일 분리
